libminkteec: Add support for Memory Object backed TEE Shared Memory
Added support for TEEC_SharedMemory to be backed by a Memory Object for
for large memory allocations (beyond a page size).

A Memory Object represents contiguous page-aligned memory shared with QTEE.
When a client requests for either allocation or registration of a Shared
Memory more than a 4 KB in size, it is associated with a Memory Object to
support the request.
TEEC_TempMemoryReference of size beyond 4KB are also supported now via an
intermediary conversion to TEEC_SharedMemory backed by a Memory Object.

Signed-off-by: Harshal Dev <quic_hdev@quicinc.com>